requirements.txt
discord.py>=2.4.0
openai>=1.40.0
google-generativeai>=0.7.2